push size_t

+++ srcs/Server.cpp
@@ -79,7 +79,7 @@ void	Server::servLoop()
 			std::cout << "poll error" << std::endl;
 			throw std::exception();
 		}
-		for (int i = 0; i < _vecPollFd.size(); i++) {
+		for (size_t i = 0; i < _vecPollFd.size(); i++) {
 			if (_vecPollFd[i].revents & POLLIN) { // on verifie si le bit de POLLIN est dans revent, comme ca on check si il y a bien des trucs a lire
 				if (_vecPollFd[i].fd = _serverSocketFd) //regarde si on a le meme file descriptor que celui du socket du serveur
 					// le client est nouveau
